yacht-mobile/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "yacht-mobile"
version = "0.1.0"
description = "Mobile-oriented Docker-compatibility shim (MVP)"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0.1"]

[project.scripts]
yacht = "yacht.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
